requirements.txt
requests
google-cloud-storage
google-cloud-firestore
Flask
python-dotenv
aiohttp
openai
pyrebase